scripts/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
transformers==4.12.5
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
redis
flask-cors
rich
aiohttp
asyncio
aiofiles
umap-learn